Consolidate all copy/paste tests in one file
